Change contrib directory to extra
The contrib name (taken from other open-source projects) seems a bit confusing. Extra will probably be clearer.
The Debian data where moved under linux directory, but are also used for ARM. As these are files used to build the package (and not vital to the functionning of duniter)
they have their place in extra.
The Debian data are used by many scripts which were not updated after debian folder move.