[mod] PoW: Raise to five zeros logs displaying matched challenge
Due to the increase of the common difficulty level, it makes sense to
increase the display of logs matched challenge to reduce amount of
generated logs in order to reduce the size of the log files.
This constant shouldn't to big otherwise small CPU smiths
wont get the hint that their node is forging.