[fix] docker: entrypoint - backup conf.json
When `/etc/duniter` is not mounted as a persistent volume, the
configuration file is lost after re-creating the instance.

To prevent this, before transitionning the configuration file from
volume '/var/lib/duniter' to volume '/etc/duniter', the entrypoint
backups it into '/var/lib/duniter'.  On restart, if the entrypoint
detects that the link to '/etc/duniter/conf.json' is dangling, it
prints a big fat warning into the logs and re-use the backuped
config file.